User inheritance to avoid `Piece` structural nightmare
