+++ osu.Game.Tests/Visual/SongSelect/BeatmapCarouselV2TestScene.cs
@@ -189,7 +189,7 @@ namespace osu.Game.Tests.Visual.SongSelect
                         .Where(p => ((ICarouselPanel)p).Item?.IsVisible == true)
                         .OrderBy(p => p.Y)
                         .ElementAt(index)
-                        .ChildrenOfType<CarouselPanelPiece>().Single()
+                        .ChildrenOfType<PanelBase>().Single()
                         .TriggerClick();
             });
         }
